enable native file dialog only for whitelisted distros
If the host's xdg-desktop-portal is too old,
it will proxy every file access through the document portal.
This breaks files with external links because FreeCAD cannot find those
files.
It also breaks saving files because FreeCAD tries to append a suffix,
which is not allowed when using the portal.

Fixes issues with older xdg-desktop-portal versions #15